Compute PC size factors
Previously used the PCA's standard deviations, however when adding the
independent part of the simulated data on, there is overlap with the
PCs and therefor this would over-do the variance in the directions of
the PCs. We instead solve for the correct amount of variance in the PCs
so that the addition of independent data ontop will give us the intended
variance in the direction of the PCs. In large-scale omics, the results
tend to be almost identical to the previous version, since PCs are
composed of a huge number of features that independently don't
contribute much at all.